+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/actions/RustOpenDefinitionOperation.java
@@ -14,7 +14,7 @@ import melnorme.lang.ide.core.operations.DaemonEnginePreferences;
 import melnorme.lang.ide.ui.LangUIMessages;
 import melnorme.lang.ide.ui.editor.EditorUtils.OpenNewEditorMode;
 import melnorme.lang.ide.ui.editor.actions.AbstractOpenElementOperation;
-import melnorme.lang.ide.ui.tools.ToolProcessRunner;
+import melnorme.lang.ide.ui.tools.ToolManagerOperationHelper;
 import melnorme.lang.tooling.ast.SourceRange;
 import melnorme.lang.tooling.ops.FindDefinitionResult;
 import melnorme.utilbox.concurrency.OperationCancellation;
@@ -49,7 +49,7 @@ public class RustOpenDefinitionOperation extends AbstractOpenElementOperation {
 		String racerPath = DaemonEnginePreferences.DAEMON_PATH.get();
 		String sdkSrcPath = RustSDKPreferences.SDK_SRC_PATH.get();
 		
-		RacerFindDefinitionOperation op = new RacerFindDefinitionOperation(new ToolProcessRunner(monitor), 
+		RacerFindDefinitionOperation op = new RacerFindDefinitionOperation(new ToolManagerOperationHelper(monitor), 
 			racerPath, sdkSrcPath, range.getOffset(), line_0, col_0, inputLoc);
 		return op.executeAndProcessOutput();
 	}