+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/text/completion/RustCompletionProposalComputer.java
@@ -13,16 +13,22 @@ package com.github.rustdt.ide.ui.text.completion;
 import melnorme.lang.ide.core.operations.DaemonEnginePreferences;
 import melnorme.lang.ide.core.operations.TimeoutProgressMonitor;
 import melnorme.lang.ide.ui.editor.actions.SourceOperationContext;
+import melnorme.lang.ide.ui.text.completion.LangCompletionProposal;
 import melnorme.lang.ide.ui.text.completion.LangCompletionProposalComputer;
-import melnorme.lang.ide.ui.tools.ToolProcessRunner;
+import melnorme.lang.ide.ui.tools.ToolManagerOperationHelper;
+import melnorme.lang.tooling.CompletionProposalKind;
+import melnorme.lang.tooling.ToolCompletionProposal;
 import melnorme.lang.tooling.completion.LangCompletionResult;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.misc.Location;
 
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.jface.text.contentassist.ICompletionProposal;
+import org.eclipse.jface.text.contentassist.IContextInformation;
 
 import com.github.rustdt.ide.core.operations.RustSDKPreferences;
+import com.github.rustdt.ide.ui.RustImageProvider;
 import com.github.rustdt.tooling.ops.RacerCompletionOperation;
 
 public class RustCompletionProposalComputer extends LangCompletionProposalComputer {
@@ -40,9 +46,28 @@ public class RustCompletionProposalComputer extends LangCompletionProposalComput
 		int col_0 = context.getInvocationColumn_0();
 		Location fileLocation = context.getEditorInputLocation();
 		
-		RacerCompletionOperation racerCompletionOp = new RacerCompletionOperation(new ToolProcessRunner(pm), 
+		RacerCompletionOperation racerCompletionOp = new RacerCompletionOperation(new ToolManagerOperationHelper(pm), 
 			racerPath, sdkSrcPath, offset, line_0, col_0, fileLocation);
 		return racerCompletionOp.executeAndProcessOutput();
+	}
+	
+	@Override
+	protected RustImageProvider getImageProvider() {
+		return new RustImageProvider();
+	}
+	
+	@Override
+	protected ICompletionProposal adaptToolProposal(ToolCompletionProposal proposal) {
+		IContextInformation ctxInfo = null; // TODO: context information
+		return new LangCompletionProposal(proposal, null, getImage(proposal), ctxInfo) {
+			@Override
+			protected String getReplaceString() {
+				if(proposal.getKind() == CompletionProposalKind.Module) {
+					return super.getReplaceString() + "::";
+				}
+				return super.getReplaceString();
+			}
+		};
 	}
 	
 }