+++ plugin_tooling/src/com/github/rustdt/tooling/ops/RacerCompletionOperation.java
@@ -11,7 +11,7 @@
 package com.github.rustdt.tooling.ops;
 
 import melnorme.lang.tooling.completion.LangCompletionResult;
-import melnorme.lang.tooling.ops.IProcessRunner;
+import melnorme.lang.tooling.ops.IOperationHelper;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.misc.Location;
@@ -21,9 +21,9 @@ public class RacerCompletionOperation extends RacerOperation {
 	
 	protected final int offset;
 	
-	public RacerCompletionOperation(IProcessRunner processRunner, String racerPath, String rustSrcPath, int offset,
+	public RacerCompletionOperation(IOperationHelper opHelper, String racerPath, String rustSrcPath, int offset,
 			int line_0, int col_0, Location fileLocation) {
-		super(processRunner, racerPath, rustSrcPath, getArguments("complete-with-snippet", line_0, col_0, fileLocation));
+		super(opHelper, racerPath, rustSrcPath, getArguments("complete-with-snippet", line_0, col_0, fileLocation));
 		
 		this.offset = offset;
 	}
@@ -31,7 +31,7 @@ public class RacerCompletionOperation extends RacerOperation {
 	public LangCompletionResult executeAndProcessOutput() throws CommonException, OperationCancellation {
 		ExternalProcessResult result = execute();
 		
-		return new RacerCompletionOutputParser(offset).parse(result);
+		return createRacerOutputParser(offset).parse(result);
 	}
 	
 }